+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enh #6: Refactor `RequestFactory` to improve performance (@Tigrov)
 - Enh #7: Apply PHP CS Fixer (@Tigrov)
 - Enh #8: Refactor `FrankenPHPApplicationRunner` to improve performance (@Tigrov)
+- Bug #9: Fix view response with failure (@Tigrov)
 
 ## 1.0.0 April 12, 2026